:recycle: [toggle_visibility_controller] Refacto stimulus controller in order to make it more generic
